+++ config/info.php
@@ -15,6 +15,6 @@ return [
 
     'eliasis' => [
 
-        'version' => '1.0.8',
+        'version' => '1.0.9',
     ],
 ];