+++ config/modules.php
@@ -6,7 +6,7 @@
  * @copyright  Copyright (c) 2017
  * @license    https://opensource.org/licenses/MIT - The MIT License (MIT)
  * @link       https://github.com/Eliasis-Framework/Eliasis
- * @since      1.0.0
+ * @since      1.0.9
  */
 
 return [